requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
WTForms==3.2.1
email-validator==2.2.0
python-dotenv==1.0.1

